requirements.txt
# ==================================================
# LIAO AI Assistant - Requirements
# ==================================================

# -------------------------------
# Backend Framework
# -------------------------------
fastapi
uvicorn

# -------------------------------
# Environment / Config
# -------------------------------
python-dotenv

# -------------------------------
# Validation / Models
# -------------------------------
pydantic

# -------------------------------
# HTTP / API Requests
# -------------------------------
requests
httpx

# -------------------------------
# Database
# -------------------------------
sqlalchemy

# -------------------------------
# AI Providers
# -------------------------------
google-genai
groq

# -------------------------------
# Desktop UI
# -------------------------------
customtkinter
pillow

# -------------------------------
# Voice Input / Output
# -------------------------------
edge-tts
speechrecognition
pyaudio

# -------------------------------
# Desktop Automation
# -------------------------------
pyautogui
psutil

# -------------------------------
# Optional Utilities
# -------------------------------
typing-extensions